refactor(metrics): extract hotspots module
Move StopHotspot struct and stop_hotspots function into a dedicated
hotspots.rs module, re-exported from metrics/mod.rs.
